requirements.txt
numpy>=1.20.3,<2.0
Flask>=2.2.2
SQLAlchemy>=2.0.20
torch>=2.0
Pillow>=8.1.0
pillow-heif>=0.14.0
imagehash>=4.3.0
transformers>=4.28.1
accelerate>=1.5.0
opencv-python-headless>=4.7.0.68
python-dotenv>=0.19.2
tqdm>=4.66.1
requests>=2.31.0
faiss-cpu>=1.11.0